# spikes/api/results/drf-openapi.json
{
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "description": "Document the normalized experimental error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorObject"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ErrorObject": {
        "description": "Document normalized safe API error fields.",
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "details",
          "message"
        ],
        "type": "object"
      },
      "Health": {
        "description": "Document the deliberately minimal health response.",
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "PatchedProjectPatchRequest": {
        "description": "Validate an explicit, non-empty Project partial update.",
        "properties": {
          "description": {
            "maxLength": 2000,
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatusEnum"
          }
        },
        "type": "object"
      },
      "Project": {
        "description": "Render the common Project response shape.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "readOnly": true,
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "readOnly": true,
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatusEnum"
          },
          "updated_at": {
            "format": "date-time",
            "readOnly": true,
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "readOnly": true,
            "type": "string"
          }
        },
        "required": [
          "created_at",
          "id",
          "name",
          "updated_at",
          "workspace_id"
        ],
        "type": "object"
      },
      "ProjectCreateRequest": {
        "description": "Validate the shared Project creation contract.",
        "properties": {
          "description": {
            "default": "",
            "maxLength": 2000,
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectStatusEnum"
              }
            ],
            "default": "active"
          },
          "workspace_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "name",
          "workspace_id"
        ],
        "type": "object"
      },
      "ProjectPage": {
        "description": "Document the deterministic list response.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "type": "array"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "type": "object"
      },
      "ProjectStatusEnum": {
        "description": "* `active` - Active\n* `archived` - Archived",
        "enum": [
          "active",
          "archived"
        ],
        "type": "string"
      }
    },
    "securitySchemes": {
      "syntheticActor": {
        "description": "Synthetic actor only. X-Workspace-ID is also required; neither header is a production authentication design.",
        "in": "header",
        "name": "X-Spike-Actor",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Experimental DRF contract; not a production API.",
    "title": "Omnilyzer API Spike - DRF",
    "version": "1.0.0-spike"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/drf/health/": {
      "get": {
        "description": "Report only a static healthy status.",
        "operationId": "drf_health",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/drf/projects/": {
      "get": {
        "description": "Validate query parameters and return a deterministic Project page.",
        "operationId": "drf_list_projects",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "* `active` - Active\n* `archived` - Archived",
            "in": "query",
            "name": "status",
            "schema": {
              "enum": [
                "active",
                "archived"
              ],
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPage"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "syntheticActor": []
          }
        ],
        "tags": [
          "projects"
        ]
      },
      "post": {
        "description": "Validate input and create a Workspace-scoped Project.",
        "operationId": "drf_create_project",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "syntheticActor": []
          }
        ],
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/drf/projects/{project_id}/": {
      "delete": {
        "description": "Delete one authorized Project and return an empty 204 response.",
        "operationId": "drf_delete_project",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No response body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "syntheticActor": []
          }
        ],
        "tags": [
          "projects"
        ]
      },
      "get": {
        "description": "Retrieve a Project without revealing cross-Workspace existence.",
        "operationId": "drf_retrieve_project",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "syntheticActor": []
          }
        ],
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "description": "Apply a validated non-empty partial update.",
        "operationId": "drf_update_project",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchedProjectPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "syntheticActor": []
          }
        ],
        "tags": [
          "projects"
        ]
      }
    }
  }
}